Fixed the image binding. Added Viewport into LightCategory, so I can put the light VP matrix in the shader instead of just the view.

+++ source/Core/Castor3D/Src/Castor3DPrerequisites.cpp
@@ -18,7 +18,7 @@ namespace Castor3D
 											 , GLSL::Vec3 & p_normal
 											 , GLSL::Float & p_shininess
 											 , uint16_t p_textureFlags
-											 , uint8_t p_programFlags
+											 , uint16_t p_programFlags
 											 , uint8_t p_sceneFlags )
 	{
 		using namespace GLSL;
@@ -50,7 +50,7 @@ namespace Castor3D
 											  , GLSL::Vec3 & p_specular
 											  , GLSL::Vec3 & p_emissive
 											  , uint16_t p_textureFlags
-											  , uint8_t p_programFlags
+											  , uint16_t p_programFlags
 											  , uint8_t p_sceneFlags )
 	{
 		using namespace GLSL;
